+++ apps/hubble/src/addon/src/store/mod.rs
@@ -1,10 +1,15 @@
+pub use self::cast_store::*;
+pub use self::link_store::*;
 pub use self::message::*;
 pub use self::reaction_store::*;
 pub use self::store::*;
 pub use self::store_event_handler::*;
 pub use self::user_data_store::*;
 pub use self::utils::*;
+pub use self::verification_store::*;
 
+mod cast_store;
+mod link_store;
 mod message;
 mod name_registry_events;
 mod reaction_store;
@@ -12,3 +17,4 @@ mod store;
 mod store_event_handler;
 mod user_data_store;
 mod utils;
+mod verification_store;