+++ apps/hubble/src/rustfunctions.ts
@@ -2,13 +2,13 @@
 // We import the Rust code as a NodeJS module, and then export it as a JS function.
 // Note that we need to use the `createRequire` function to import the module, since it
 // is binary code. If we used `import` instead, it would be interpreted as a JS module, and
-// we would get an error becaues it would try to parse it as JS
+// we would get an error because it would try to parse it as JS
 import { createRequire } from "module";
 const require = createRequire(import.meta.url);
 const lib = require("./addon/index.node");
 
-import { HubError, HubErrorCode, validations } from "@farcaster/hub-nodejs";
-import { PAGE_SIZE_MAX, PageOptions } from "./storage/stores/types.js";
+import { HubError, HubErrorCode, LinkAddMessage, validations } from "@farcaster/hub-nodejs";
+import { MessagesPage, PAGE_SIZE_MAX, PageOptions } from "./storage/stores/types.js";
 import { UserMessagePostfix } from "./storage/db/types.js";
 import { DbKeyValue, RocksDbIteratorOptions } from "./storage/db/rocksdb.js";
 import { logger } from "./utils/logger.js";
@@ -69,16 +69,16 @@ export const rsCreateStatsdClient = (host: string, port: number, prefix: string)
   lib.createStatsdClient(host, port, prefix);
 };
 
-/** Create or Open a DB at a give path 
- * 
+/** Create or Open a DB at a give path
+ *
  * All rust objects need to be "owned" by someone so that rust can manage its lifecycle. For rust objects like the
- *  `RocksDb` and the `ReactionStore`, we create them as `JsBox<Arc<T>>`. `JsBox` is like Rust's `Box`, except it is 
- * owned by the Javascript pointer that is returned. That is, 
+ *  `RocksDb` and the `ReactionStore`, we create them as `JsBox<Arc<T>>`. `JsBox` is like Rust's `Box`, except it is
+ * owned by the Javascript pointer that is returned. That is,
     - The DB object is owned by the Javascript object that is returned
-    - When the Javascript object goes out of scope and is gc'd, it is `drop()`-ed in Rust. 
+    - When the Javascript object goes out of scope and is gc'd, it is `drop()`-ed in Rust.
 
-  Since the Rust objects are `Arc<T>` inside a `JsBox`, we can clone them and keep them around in the rust code as we 
-  please, since the Javascript code will continue to own one `Arc<T>`, making sure that it lasts for the lifetime of 
+  Since the Rust objects are `Arc<T>` inside a `JsBox`, we can clone them and keep them around in the rust code as we
+  please, since the Javascript code will continue to own one `Arc<T>`, making sure that it lasts for the lifetime of
   the program.
 */
 export const rsCreateDb = (path: string): RustDb => {
@@ -142,30 +142,30 @@ export const rsDbCommit = async (db: RustDb, keyValues: DbKeyValue[]): Promise<v
 };
 
 /**
- * Rust code needs to be memory-safe, which means that we can't pass around iterators like we do in Javascript. 
- * This is because the `iterator` reference is valid for only as long as the `db` is valid, and the reference is 
+ * Rust code needs to be memory-safe, which means that we can't pass around iterators like we do in Javascript.
+ * This is because the `iterator` reference is valid for only as long as the `db` is valid, and the reference is
  * dropped right after the iterator is finished.
 
   This specifically means that we need to use iterators as callbacks. The way the iterators are set up is:
   - Call the `forEachIteartor` method with your callback (Either in JS or Rust)
   - Perform all actions in the callback
-  - At the end of the itearation, the iterator is returned and closed by Rust
+  - At the end of the iteration, the iterator is returned and closed by Rust
 
   In JS, we can have async functions as callbacks to the `forEachIterator` methods. This means that the callback
   can take arbitrarily long, and that is bad because keeping iterators open for long periods of time is very
-  problematic. Additionally, we can't call async JS methods from rust. To address these both, the iterators are 
-  automatically paged. 
+  problematic. Additionally, we can't call async JS methods from rust. To address these both, the iterators are
+  automatically paged.
 
   That means that when you start an iterator:
   1. JS code will fetch a page full of keys and values from rust
-  2. Close the iterator right after. 
-  3. Calls the async callbacks with the cached key, value parirs, which can take as long as needed. 
-  4. Go back to step 1 to get the next page of key, value pairs. 
+  2. Close the iterator right after.
+  3. Calls the async callbacks with the cached key, value pairs, which can take as long as needed.
+  4. Go back to step 1 to get the next page of key, value pairs.
 
-  This method returns a boolean, which is true if the iteration is finished, and false if it is not. 
+  This method returns a boolean, which is true if the iteration is finished, and false if it is not.
   - If the iteration was stopped because it hit the pageSize, it returns false (i.e., there are more keys available)
   - If the iteration was stopped because the callback returned true, it returns false (i.e., there are more keys available)
-  
+
  */
 export const rsDbForEachIteratorByPrefix = async (
   db: RustDb,
@@ -305,6 +305,58 @@ export const rsCreateReactionStore = (
   return store as RustDynStore;
 };
 
+/** Create a cast Store */
+export const rsCreateCastStore = (
+  db: RustDb,
+  eventHandler: RustStoreEventHandler,
+  pruneSizeLimit: number,
+): RustDynStore => {
+  const store = lib.createCastStore(db, eventHandler, pruneSizeLimit);
+
+  return store as RustDynStore;
+};
+
+export const rsGetCastAdd = async (store: RustDynStore, fid: number, hashBytes: Buffer): Promise<Buffer> => {
+  return await lib.getCastAdd.call(store, fid, hashBytes);
+};
+
+export const rsGetCastRemove = async (store: RustDynStore, fid: number, hashBytes: Buffer): Promise<Buffer> => {
+  return await lib.getCastRemove.call(store, fid, hashBytes);
+};
+
+export const rsGetCastAddsByFid = async (
+  store: RustDynStore,
+  fid: number,
+  pageOptions: PageOptions,
+): Promise<RustMessagesPage> => {
+  return await lib.getCastAddsByFid.call(store, fid, pageOptions);
+};
+
+export const rsGetCastRemovesByFid = async (
+  store: RustDynStore,
+  fid: number,
+  pageOptions: PageOptions,
+): Promise<RustMessagesPage> => {
+  return await lib.getCastRemovesByFid.call(store, fid, pageOptions);
+};
+
+export const rsGetCastsByParent = async (
+  store: RustDynStore,
+  parentCastIdBytes: Buffer,
+  parentUrl: string,
+  pageOptions: PageOptions,
+): Promise<RustMessagesPage> => {
+  return await lib.getCastsByParent.call(store, parentCastIdBytes, parentUrl, pageOptions);
+};
+
+export const rsGetCastsByMention = async (
+  store: RustDynStore,
+  mentionFid: number,
+  pageOptions: PageOptions,
+): Promise<RustMessagesPage> => {
+  return await lib.getCastsByMention.call(store, mentionFid, pageOptions);
+};
+
 export const rsGetMessage = async (
   store: RustDynStore,
   fid: number,
@@ -390,6 +442,7 @@ export const rsGetReactionsByTarget = async (
   return await lib.getReactionsByTarget.call(store, targetCastIdBytes, targetUrl, type, pageOptions);
 };
 
+/** UserData Store */
 export const rsCreateUserDataStore = (
   db: RustDb,
   eventHandler: RustStoreEventHandler,
@@ -423,3 +476,105 @@ export const rsGetUserNameProofByFid = async (store: RustDynStore, fid: number):
 export const rsMergeUserNameProof = async (store: RustDynStore, usernameProof: Uint8Array): Promise<Buffer> => {
   return await lib.mergeUserNameProof.call(store, usernameProof);
 };
+
+/** VerificationStore */
+export const rsCreateVerificationStore = (
+  db: RustDb,
+  eventHandler: RustStoreEventHandler,
+  pruneSizeLimit: number,
+): RustDynStore => {
+  const store = lib.createVerificationStore(db, eventHandler, pruneSizeLimit);
+
+  return store as RustDynStore;
+};
+
+export const rsGetVerificationAdd = async (store: RustDynStore, fid: number, address: Uint8Array): Promise<Buffer> => {
+  return await lib.getVerificationAdd.call(store, fid, address);
+};
+
+export const rsGetVerificationRemove = async (
+  store: RustDynStore,
+  fid: number,
+  address: Uint8Array,
+): Promise<Buffer> => {
+  return await lib.getVerificationRemove.call(store, fid, address);
+};
+
+export const rsGetVerificationAddsByFid = async (
+  store: RustDynStore,
+  fid: number,
+  pageOptions: PageOptions,
+): Promise<RustMessagesPage> => {
+  return await lib.getVerificationAddsByFid.call(store, fid, pageOptions);
+};
+
+export const rsGetVerificationRemovesByFid = async (
+  store: RustDynStore,
+  fid: number,
+  pageOptions: PageOptions,
+): Promise<RustMessagesPage> => {
+  return await lib.getVerificationRemovesByFid.call(store, fid, pageOptions);
+};
+
+export const rsMigrateVerifications = async (store: RustDynStore): Promise<{ total: number; duplicates: number }> => {
+  return await lib.migrateVerifications.call(store);
+};
+export namespace rsLinkStore {
+  export const CreateLinkStore = (
+    db: RustDb,
+    eventHandler: RustStoreEventHandler,
+    pruneSizeLimit: number,
+  ): RustDynStore => {
+    const store = lib.createLinkStore(db, eventHandler, pruneSizeLimit);
+
+    return store as RustDynStore;
+  };
+
+  export const GetLinkAdd = async (store: RustDynStore, fid: number, type: string, target: number): Promise<Buffer> => {
+    return await lib.getLinkAdd.call(store, fid, type, target);
+  };
+
+  export const GetLinkAddsByFid = async (
+    store: RustDynStore,
+    fid: number,
+    type: string,
+    pageOptions: PageOptions,
+  ): Promise<RustMessagesPage> => {
+    return await lib.getLinkAddsByFid.call(store, fid, type, pageOptions);
+  };
+
+  export const GetLinkRemovesByFid = async (
+    store: RustDynStore,
+    fid: number,
+    type: string,
+    pageOptions: PageOptions,
+  ): Promise<RustMessagesPage> => {
+    return await lib.getLinkRemovesByFid.call(store, fid, type, pageOptions);
+  };
+
+  export const GetLinksByTarget = async (
+    store: RustDynStore,
+    target: number,
+    type: string,
+    pageOptions: PageOptions,
+  ): Promise<RustMessagesPage> => {
+    return await lib.getLinksByTarget.call(store, target, type, pageOptions);
+  };
+
+  export const GetLinkRemove = async (
+    store: RustDynStore,
+    fid: number,
+    type: string,
+    target: number,
+  ): Promise<Buffer> => {
+    return await lib.getLinkRemove.call(store, fid, type, target);
+  };
+
+  export const GetAllLinkMessagesByFid = async (
+    store: RustDynStore,
+    fid: number,
+    pageOptions: PageOptions,
+  ): Promise<RustMessagesPage> => {
+    return await lib.getAllLinkMessagesByFid.call(store, fid, pageOptions);
+  };
+}