+++ apps/hubble/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @farcaster/hubble
 
+## 1.10.11
+
+### Patch Changes
+
+- 22615b3c: Add address to frame message
+- Updated dependencies [22615b3c]
+  - @farcaster/hub-nodejs@0.11.7
+
 ## 1.10.10
 
 ### Patch Changes